fix connection reliability

+++ controlbox/src/cbox/DataStream.h
@@ -26,7 +26,7 @@
 #include <string>
 namespace cbox {
 
-using stream_size_t = uint16_t;
+using stream_size_t = size_t;
 
 /**
  * An output stream that supports writing data.
